refactor: commonize PlainMessageSharedKey*
