+++ serde-encrypt-core/src/encrypt.rs
@@ -3,4 +3,3 @@
 pub mod encrypted_message;
 pub mod plain_message_public_key;
 pub mod plain_message_shared_key;
-pub mod plain_message_shared_key_deterministic;
